MB-45 api response dtos done

+++ src/main/java/com/mrivals_builder/Mrivals_Builder/entities/LeaderboardPlayer.java
@@ -19,7 +19,7 @@ public class LeaderboardPlayer {
 
     private double rankScore;
     private double maxRankScore;
-    private double seasonWinCount;
+    private int seasonWinCount;
 
     @ManyToOne
     private Hero hero;
